+++ LineCircleIntersection/LineCircleIntersection.cs
@@ -48,7 +48,7 @@ namespace LineCircleIntersection
 		protected void DrawSphere(Sphere3D sphere, Graphics graphics, Color color)
 		{
 			// Create point bounds
-			RectangleF pointBounds = new RectangleF(sphere.Center.X - sphere.Radius, sphere.Center.Y - sphere.Radius, 2f * sphere.Radius, 2f * sphere.Radius);
+			RectangleF pointBounds = new RectangleF((float)sphere.Center.X - (float)sphere.Radius, (float)sphere.Center.Y - (float)sphere.Radius, 2f * (float)sphere.Radius, 2f * (float)sphere.Radius);
 
 			// Create Pen
 			Brush brush = new SolidBrush(color);
@@ -67,7 +67,7 @@ namespace LineCircleIntersection
 			const float width = 6f;
 			
 			// Create point bounds
-			RectangleF pointBounds = new RectangleF(point.X - 0.5f * width, point.Y - 0.5f * width, width, width);
+			RectangleF pointBounds = new RectangleF((float)point.X - 0.5f * width, (float)point.Y - 0.5f * width, width, width);
 
 			// Create Pen
 			Brush brush = new SolidBrush(color);
@@ -86,8 +86,8 @@ namespace LineCircleIntersection
 			Pen pen = new Pen(color, width);
 
 			// Create Drawing.Points
-			PointF pt1 = new PointF(start.X, start.Y);
-			PointF pt2 = new PointF(end.X, end.Y);
+			PointF pt1 = new PointF((float)start.X, (float)start.Y);
+			PointF pt2 = new PointF((float)end.X, (float)end.Y);
 
 			// Draw point
 			graphics.DrawLine(pen, pt1, pt2);
@@ -99,8 +99,8 @@ namespace LineCircleIntersection
 			Pen pen = new Pen(color, width);
 
 			// Create Drawing.Points
-			PointF pt1 = new PointF(line.Start.X, line.Start.Y);
-			PointF pt2 = new PointF(line.End.X, line.End.Y);
+			PointF pt1 = new PointF((float)line.Start.X, (float)line.Start.Y);
+			PointF pt2 = new PointF((float)line.End.X, (float)line.End.Y);
 
 			// Draw point
 			graphics.DrawLine(pen, pt1, pt2);
@@ -117,8 +117,8 @@ namespace LineCircleIntersection
 
 			// Create Drawing.Points
 			Vector3D endPoint = line.Origin + line.Direction * lineSegment3D.Length();
-			PointF pt1 = new PointF(line.Origin.X, line.Origin.Y);
-			PointF pt2 = new PointF(endPoint.X, endPoint.Y);
+			PointF pt1 = new PointF((float)line.Origin.X, (float)line.Origin.Y);
+			PointF pt2 = new PointF((float)endPoint.X, (float)endPoint.Y);
 
 			// Draw point
 			graphics.DrawLine(pen, pt1, pt2);
